+++ src/class/frontend/link_makers/class-link-factory.php
@@ -23,7 +23,6 @@ use Lumiere\Link_Makers\No_Links;
 use Lumiere\Link_Makers\Bootstrap_Links;
 use Lumiere\Link_Makers\Highslide_Links;
 use Lumiere\Link_Makers\Classic_Links;
-use Lumiere\Tools\Utils;
 use Lumiere\Tools\Settings_Global;
 
 /**
@@ -31,7 +30,9 @@ use Lumiere\Tools\Settings_Global;
  */
 class Link_Factory {
 
-	// Trait including the database settings.
+	/**
+	 * Traits
+	 */
 	use Settings_Global;
 
 	/**
@@ -53,7 +54,7 @@ class Link_Factory {
 		/**
 		 * Checks if the current page is AMP
 		 */
-		if ( Utils::lumiere_is_amp_page() === true ) {
+		if ( $this->lumiere_is_amp_page() === true ) {
 			return new AMP_Links();
 		}
 
@@ -97,5 +98,46 @@ class Link_Factory {
 
 		return ( new self() )->lumiere_select_link_maker();
 
+	}
+
+	/**
+	 * Are we currently on an AMP URL?
+	 * Always return `false` and show PHP Notice if called before the `wp` hook.
+	 *
+	 * @since 3.7.1
+	 * @return bool true if amp url, false otherwise
+	 */
+	private function lumiere_is_amp_page(): bool {
+
+		global $pagenow;
+
+		// If url contains ?amp, it must be an AMP page
+		if ( str_contains( $_SERVER['REQUEST_URI'] ?? '', '?amp' )
+		|| isset( $_GET ['wpamp'] )
+		|| isset( $_GET ['amp'] )
+		) {
+			return true;
+		}
+
+		if (
+			is_admin()
+			/**
+			 * If kept, breaks blog pages these functions can be executed very early
+				|| is_embed()
+				|| is_feed()
+			*/
+			|| ( isset( $pagenow ) && in_array( $pagenow, [ 'wp-login.php', 'wp-signup.php', 'wp-activate.php' ], true ) )
+			|| ( defined( 'REST_REQUEST' ) && REST_REQUEST )
+			|| ( defined( 'XMLRPC_REQUEST' ) && XMLRPC_REQUEST )
+		) {
+			return false;
+		}
+
+		// Since we are checking later (amp_is_request()) a function that execute late, make sure we can execute it
+		if ( did_action( 'wp' ) === 0 ) {
+			return false;
+		}
+
+		return function_exists( 'amp_is_request' ) && amp_is_request();
 	}
 }